// frontend/src/api/generated/types.gen.ts
export type AuditLogApplicationIdsResponse = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    applicationIds: Array<string>;
};

export type AuditLogClientIdsResponse = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    clientIds: Array<string>;
};

export type AuditLogEntityTypesResponse = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    entityTypes: Array<string>;
};

export type AuditLogListResponse = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    auditLogs: Array<AuditLogResponse>;
    hasMore: boolean;
    nextCursor?: string;
};

export type AuditLogOperationsResponse = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    operations: Array<string>;
};

export type AuditLogResponse = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    applicationId?: string;
    clientId?: string;
    entityId: string;
    entityType: string;
    id: string;
    operation: string;
    operationJson?: string;
    performedAt: string;
    principalId?: string;
    principalName?: string;
};